feat: update authentication slice and validation logic, refactor header styles, and improve password requirements
