+++ src/components/header/Logo.tsx
@@ -1,6 +1,6 @@
 import { NavLink } from "react-router-dom";
 import { Heading } from "../common";
-import styles from "../../styles/components/common/Header.module.scss";
+import styles from "../../styles/components/header/Header.module.scss";
 import clsx from "clsx";
 
 type LogoProps = {
